+++ lib/browser_core.js
@@ -3,7 +3,7 @@
 const path = require('path');
 const ErrorFactory = require('./errors/error_factory');
 const config = require('../config');
-const DomElement = require('./dom_element');
+const DomElement = require('./models/dom_element');
 
 const injectionScriptsPath = config.injectionScripts.path;
 const injectionScripts = config.injectionScripts.files;
