+++ lib/mixins/browser_actions.js
@@ -2,7 +2,7 @@
 "use strict";
 
 const ErrorFactory = require('../errors/error_factory');
-const DomElement = require('../dom_element');
+const DomElement = require('../models/dom_element');
 const utils = require('../utils');
 
 module.exports = function BrowserActionsMixin(s) {
