+++ lib/mixins/browser_queries.js
@@ -1,6 +1,6 @@
 "use strict";
 
-const DomElement = require('../dom_element');
+const DomElement = require('../models/dom_element');
 const ErrorFactory = require('../errors/error_factory');
 
 module.exports = function BrowserQueriesMixin(s) {
