+++ lib/models/dom_element.js
